requirements.txt
numpy
wandb
pytorch-lightning
scikit-image
pandas
pyarrow
imageio
matplotlib
torchmetrics[image]
hydra-core
opencv-python
Pillow
jaxtyping
nibabel
beartype
blobfile
tqdm
pyYaml
pillow
einops
lpips
